Update README semantic HTML / CSS emphasis
We are looking to change the project description in a way that strongly suggests the applicant not use any CSS frameworks, along with emphasizing accessibility and mobile / responsive design considerations. We have had several submissions where it was difficult to gauge the semantic HTML / CSS skill level, so this will hopefullly help test project expectations clearer around these topics.